refactored blockSnapshots. added cy.assertFrontendStyles and chainable command
