+++ cypress/support/gutenberg/internals/index.js
@@ -1,0 +1,7 @@
+export {
+	blockSnapshotsAssertHtmlTag,
+	blockSnapshotsAssertClassName,
+	blockSnapshotsAssertHtmlAttribute,
+	blockSnapshotsAssertBlockContent,
+	blockSnapshotsAssertComputedStyle,
+} from './block-snapshots'
