--- cypress/support/gutenberg/plugins/index.js
@@ -1,3 +1,0 @@
-export {
-	registerBlockSnapshots,
-} from './blockSnapshots'
